RadixHashJoin P3: deferred exact key+ref scatter (BuildStore) with multi-target NT
Implements the build half of the radix hash join as a standalone `BuildStore`
(accumulate+select, merge, and the deferred exact key+ref scatter). Wiring into
the join executor is deferred to P5.

Scatter / partitioning:
- Stores the 32-bit `IColumn::computeHashInto` hash per row (not a derived
  `uint16` partition id); routing keys are `UInt32` across `Scatter` and
  `BuildStore`. `MAX_LEAVES` raised to 2^20.
- Accepts multi-column fixed-width keys, packed row-major with compile-time
  width specialization (no runtime `memcpy`). Chunked, cursor-based incremental
  scatter.
- Per-thread histograms + static per-thread scatter ownership via a
  per-(thread,partition) offset matrix: O(T * num_leaves), not
  O(blocks * num_leaves). Lock-free multi-threaded build.
- Depth-first multi-pass refine with incremental partition freeing.
- `GrowingArena`: non-THP, `mmap`-backed bump allocator with power-of-two
  growth and `MADV_DONTNEED` `trim`/`releaseRange`. Replaces `HugeArena`.
- `runOnPool` schedules workers with `std::packaged_task`/`std::future`. The
  replicated histogram is folded into `BuildStore` (`Selector` removed).

Non-temporal SWWC scatter:
- Unifies the v3/v4 NT kernels into one `DECLARE_MULTITARGET_CODE` block using
  `__builtin_nontemporal_store` and a `LINE_BYTES` `VecNTLine`. Built at
  `x86-64-v3`, so multi-target dispatch activates the AVX-512/AVX2 NT stores at
  runtime.
- Adds an incremental head-peel SWWC scatter (`scatterColumnIntoSwwc` /
  `scatterColumnDrainSwwc`) so the single-pass / multi-pass pass-0 worker can use
  NT even though per-thread sub-ranges of the shared contiguous leaf arrays are
  not 64 B-aligned: each (worker,partition) writes its unaligned prefix directly
  until aligned, then NT-flushes whole lines and drains the residual at the end.
  Worker-boundary cache lines are co-owned only by byte-exact direct stores; a
  per-worker fence makes the NT stores visible before the pool join.
- `shouldUseSwwc` engages SWWC at fanout >= 256 (the measured crossover). The NT
  scatter primitive wins ~1.2-1.5x over direct at fanout >= 256
  (`bench_radix_sweep_native`); the integrated build-scatter is perf-neutral
  (per-row head-peel overhead), kept as future-proofing.

`BuildRef{0,0}` zero-init empty sentinel + 1-based `row_no`:
- `row_no` in `BuildRef` is now **1-based**: build row `r` in block `b` is
  stored as `BuildRef{b, r+1}`. `row_no == 0` is reserved as the leaf-cell
  empty sentinel, making the all-zero `BuildRef{0,0}` the empty state.
- The leaf-HT cell array is carved from the `mmap(MAP_ANONYMOUS)`-backed
  `GrowingArena`, which zero-initialises pages on first touch; a fresh leaf HT
  is therefore already empty with no `memset(0xFF)` init pass needed (spec §5.6).
- Payload is resolved at probe time with `row_no - 1`. The `verifyConservationAndRefs`
  gtest asserts `row_no >= 1` and decodes via `row = row_no - 1`.

Tests: conservation / per-leaf membership / routing identity across single-pass,
multi-pass, multi-column, large-leaf (>65536), and single-pass + multi-pass NT
SWWC; `MemoryConsumptionTest`; the `bench_radix_sweep_native` NT-vs-direct
calibration harness. The NT kernel is byte-for-byte equal to the scalar
reference; gtests and clang-tidy are clean.

+++ src/Common/tests/bench_radix_sweep_native.cpp
@@ -20,10 +20,11 @@
 //
 // "nt/bt" = direct/swwc (> 1 means write-combining + NT still wins after alloc+fault is counted).
 //
-// NOTE: SWWC exists only when NT stores are available. In the default x86-64-v2 build
-// (ENABLE_MULTITARGET_CODE=0) NT is dormant, so scatterColumn(use_swwc=true) runs the DIRECT path and
-// the "swwc" and "direct" columns are identical. The genuine NT path (vmovntdq/vmovntps) is only
-// emitted in a multitarget build, where SWWC+NT wins at high fanout (P >= 2048). See
+// NOTE: SWWC exists only when NT stores are available. This build is `x86-64-v3`
+// (`ENABLE_MULTITARGET_CODE=1`), so NT is active: `scatterColumn(use_swwc=true)` runs the genuine NT
+// path (`vmovntdq`/`vmovntps`) and the "swwc" and "direct" columns differ, with SWWC+NT winning at
+// high fanout (P >= 2048). In a `x86-64-v2` build (`ENABLE_MULTITARGET_CODE=0`) NT is dormant, so
+// `scatterColumn(use_swwc=true)` falls back to the DIRECT path and the two columns are identical. See
 // analysis/radix_hash_p2_analysis.md. (Key widths swept here are all multiples of 4, the supported set.)
 
 #include <pthread.h>
@@ -147,7 +148,7 @@ double measure(
     int P,
     int T,
     size_t rows_per_thread,
-    const std::vector<UInt16> & pid,
+    const std::vector<UInt32> & pid,
     const std::vector<char> & keybuf,
     const std::vector<BuildRef> & refs,
     bool swwc)
@@ -227,7 +228,7 @@ int main(int /*argc*/, char ** /*argv*/)
 
     fmt::print("Generating {} rows (pid 14-bit + {} GiB key bytes + 8 B ref)...\n", kTotalRows, kKeyBufBytes >> 30);
     const auto tg0 = Clk::now();
-    std::vector<UInt16> pid(kTotalRows);
+    std::vector<UInt32> pid(kTotalRows);
     std::vector<char> keybuf(kKeyBufBytes);
     std::vector<BuildRef> refs(kTotalRows);
     {
@@ -236,7 +237,7 @@ int main(int /*argc*/, char ** /*argv*/)
         for (size_t i = 0; i < kTotalRows; ++i)
         {
             const UInt64 k = rng();
-            pid[i] = static_cast<UInt16>(k & pid_mask); /// uniform in [0, 16384)
+            pid[i] = static_cast<UInt32>(k & pid_mask); /// uniform in [0, 16384)
             refs[i] = BuildRef{0, static_cast<UInt32>(i)};
         }
         /// Key content is irrelevant to throughput; fill the reusable key buffer with a cheap pattern.
